feat: add fetch_run_exports and generalize all the code
