--- crates/rattler_gateway/Cargo.toml
@@ -1,94 +1,0 @@
-[package]
-name = "repodata_gateway"
-version.workspace = true
-edition.workspace = true
-categories.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-readme.workspace = true
-
-[dependencies]
-async-compression = { version = "0.4.3", features = [
-    "gzip",
-    "tokio",
-    "bzip2",
-    "zstd",
-] }
-blake2 = "0.10.6"
-cache_control = "0.2.0"
-chrono = { version = "0.4.31", default-features = false, features = [
-    "std",
-    "serde",
-    "alloc",
-    "clock",
-] }
-humansize = "2.1.3"
-humantime = "2.1.0"
-futures = "0.3.28"
-reqwest = { version = "0.11.22", default-features = false, features = [
-    "stream",
-] }
-reqwest-middleware = "0.2.4"
-tokio-util = { version = "0.7.9", features = ["codec", "io"] }
-tempfile = "3.8.0"
-tracing = "0.1.37"
-thiserror = "1.0.49"
-url = { version = "2.4.1", features = ["serde"] }
-tokio = { version = "1.32.0", features = ["rt", "io-util"] }
-anyhow = "1.0.75"
-serde = { version = "1.0.188", features = ["derive"] }
-serde_json = { version = "1.0.107" }
-pin-project-lite = "0.2.13"
-md-5 = "0.10.6"
-rattler_digest = { version = "0.16.2", path = "../rattler_digest", features = [
-    "tokio",
-    "serde",
-] }
-rattler_conda_types = { version = "0.16.2", path = "../rattler_conda_types", optional = true }
-fxhash = { version = "0.2.1", optional = true }
-memmap2 = { version = "0.7.1", optional = true }
-ouroboros = { version = "0.17.2", optional = true }
-serde_with = "3.3.0"
-superslice = { version = "1.0.0", optional = true }
-itertools = { version = "0.11.0", optional = true }
-json-patch = "1.1.0"
-hex = { version = "0.4.3", features = ["serde"] }
-rattler_networking = { version = "0.16.2", path = "../rattler_networking", default-features = false }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-
-[target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.48.0", features = [
-    "Win32_Storage_FileSystem",
-    "Win32_Foundation",
-    "Win32_System_IO",
-] }
-
-[dev-dependencies]
-hex-literal = "0.4.1"
-tower-http = { version = "0.4.4", features = [
-    "fs",
-    "compression-gzip",
-    "trace",
-] }
-tracing-test = { version = "0.2.4" }
-insta = { version = "1.33.0", features = ["yaml"] }
-axum = "0.6.20"
-assert_matches = "1.5.0"
-tokio = { version = "1.32.0", features = ["macros", "rt-multi-thread"] }
-rstest = "0.18.2"
-
-[features]
-default = ['native-tls', 'sparse']
-native-tls = ['reqwest/native-tls']
-rustls-tls = ['reqwest/rustls-tls']
-sparse = [
-    "rattler_conda_types",
-    "memmap2",
-    "ouroboros",
-    "superslice",
-    "itertools",
-    "serde_json/raw_value",
-]
