--- crates/rattler_gateway/src/lib.rs
@@ -1,9 +1,0 @@
-#![deny(missing_docs)]
-
-//! TODO: docs
-
-pub mod fetch;
-#[cfg(feature = "sparse")]
-pub mod sparse;
-
-mod utils;
