+++ crates/rattler_repodata_gateway/src/sparse/mod.rs
@@ -1,32 +1,27 @@
-//! This module provides the [`SparseRepoData`] which is a struct to enable only sparsely loading records
-//! from a `repodata.json` file.
+//! Sparse data parsing for conda package information such as, `repodata.json`, and `run_exports.json` files.
 
 #![allow(clippy::mem_forget)]
 
-use futures::{stream, StreamExt, TryFutureExt, TryStreamExt};
 use itertools::Itertools;
-use rattler_conda_types::{
-    compute_package_url, Channel, ChannelInfo, PackageName, PackageRecord, RepoDataRecord,
-};
+use rattler_conda_types::{Channel, ChannelInfo, PackageRecord};
 use serde::{
     de::{Error, MapAccess, Visitor},
     Deserialize, Deserializer,
 };
 use serde_json::value::RawValue;
-use std::{
-    collections::{HashSet, VecDeque},
-    fmt, io,
-    marker::PhantomData,
-    path::Path,
-};
-use superslice::Ext;
+use std::{fmt, io, marker::PhantomData, path::Path};
 
-/// A struct to enable loading records from a `repodata.json` file on demand. Since most of the time you
-/// don't need all the records from the `repodata.json` this can help provide some significant speedups.
-pub struct SparseRepoData {
+pub mod repodata;
+pub mod run_exports;
+
+pub use repodata::*;
+pub use run_exports::*;
+
+/// Generalized struct for storing and loading data from large files using sparse data parsing.
+pub struct SparseData<T: Sized> {
     /// Data structure that holds a memory mapped repodata.json file and an index into the the records
     /// store in that data.
-    inner: SparseRepoDataInner,
+    inner: SparseDataInner,
 
     /// The channel from which this data was downloaded.
     channel: Channel,
@@ -36,14 +31,14 @@ pub struct SparseRepoData {
 
     /// A function that can be used to patch the package record after it has been parsed.
     /// This is mainly used to add `pip` to `python` if desired
-    patch_record_fn: Option<fn(&mut PackageRecord)>,
+    patch_function: Option<fn(&mut T)>,
 }
 
 /// A struct that holds a memory map of a `repodata.json` file and also a self-referential field which
 /// indexes the data in the memory map with a sparsely parsed json struct. See [`LazyRepoData`].
 
 #[ouroboros::self_referencing]
-struct SparseRepoDataInner {
+struct SparseDataInner {
     /// Memory map of the `repodata.json` file
     memory_map: memmap2::Mmap,
 
@@ -51,10 +46,10 @@ struct SparseRepoDataInner {
     /// map so we have to use ouroboros to make this legal.
     #[borrows(memory_map)]
     #[covariant]
-    repo_data: LazyRepoData<'this>,
+    data: LazyMap<'this>,
 }
 
-impl SparseRepoData {
+impl<T: Sized> SparseData<T> {
     /// Construct an instance of self from a file on disk and a [`Channel`].
     /// The `patch_function` can be used to patch the package record after it has been parsed
     /// (e.g. to add `pip` to `python`).
@@ -62,19 +57,19 @@ impl SparseRepoData {
         channel: Channel,
         subdir: impl Into<String>,
         path: impl AsRef<Path>,
-        patch_function: Option<fn(&mut PackageRecord)>,
+        patch_function: Option<fn(&mut T)>,
     ) -> Result<Self, io::Error> {
         let file = std::fs::File::open(path)?;
         let memory_map = unsafe { memmap2::Mmap::map(&file) }?;
-        Ok(SparseRepoData {
-            inner: SparseRepoDataInnerTryBuilder {
+        Ok(SparseData {
+            inner: SparseDataInnerTryBuilder {
                 memory_map,
-                repo_data_builder: |memory_map| serde_json::from_slice(memory_map.as_ref()),
+                data_builder: |memory_map| serde_json::from_slice(memory_map.as_ref()),
             }
             .try_build()?,
             subdir: subdir.into(),
             channel,
-            patch_record_fn: patch_function,
+            patch_function,
         })
     }
 
@@ -83,107 +78,13 @@ impl SparseRepoData {
     /// This works by iterating over all elements in the `packages` and `conda_packages` fields of
     /// the repodata and returning the unique package names.
     pub fn package_names(&self) -> impl Iterator<Item = &'_ str> + '_ {
-        let repo_data = self.inner.borrow_repo_data();
+        let repo_data = self.inner.borrow_data();
         repo_data
             .packages
             .iter()
             .chain(repo_data.conda_packages.iter())
             .map(|(name, _)| name.package)
             .dedup()
-    }
-
-    /// Returns all the records for the specified package name.
-    pub fn load_records(&self, package_name: &PackageName) -> io::Result<Vec<RepoDataRecord>> {
-        let repo_data = self.inner.borrow_repo_data();
-        let base_url = repo_data.info.as_ref().and_then(|i| i.base_url.as_deref());
-        let mut records = parse_records(
-            package_name,
-            &repo_data.packages,
-            base_url,
-            &self.channel,
-            &self.subdir,
-            self.patch_record_fn,
-        )?;
-        let mut conda_records = parse_records(
-            package_name,
-            &repo_data.conda_packages,
-            base_url,
-            &self.channel,
-            &self.subdir,
-            self.patch_record_fn,
-        )?;
-        records.append(&mut conda_records);
-        Ok(records)
-    }
-
-    /// Given a set of [`SparseRepoData`]s load all the records for the packages with the specified
-    /// names and all the packages these records depend on.
-    ///
-    /// This will parse the records for the specified packages as well as all the packages these records
-    /// depend on.
-    ///
-    pub fn load_records_recursive<'a>(
-        repo_data: impl IntoIterator<Item = &'a SparseRepoData>,
-        package_names: impl IntoIterator<Item = PackageName>,
-        patch_function: Option<fn(&mut PackageRecord)>,
-    ) -> io::Result<Vec<Vec<RepoDataRecord>>> {
-        let repo_data: Vec<_> = repo_data.into_iter().collect();
-
-        // Construct the result map
-        let mut result: Vec<_> = (0..repo_data.len()).map(|_| Vec::new()).collect();
-
-        // Construct a set of packages that we have seen and have been added to the pending list.
-        let mut seen: HashSet<PackageName> = package_names.into_iter().collect();
-
-        // Construct a queue to store packages in that still need to be processed
-        let mut pending: VecDeque<_> = seen.iter().cloned().collect();
-
-        // Iterate over the list of packages that still need to be processed.
-        while let Some(next_package) = pending.pop_front() {
-            for (i, repo_data) in repo_data.iter().enumerate() {
-                let repo_data_packages = repo_data.inner.borrow_repo_data();
-                let base_url = repo_data_packages
-                    .info
-                    .as_ref()
-                    .and_then(|i| i.base_url.as_deref());
-
-                // Get all records from the repodata
-                let mut records = parse_records(
-                    &next_package,
-                    &repo_data_packages.packages,
-                    base_url,
-                    &repo_data.channel,
-                    &repo_data.subdir,
-                    patch_function,
-                )?;
-                let mut conda_records = parse_records(
-                    &next_package,
-                    &repo_data_packages.conda_packages,
-                    base_url,
-                    &repo_data.channel,
-                    &repo_data.subdir,
-                    patch_function,
-                )?;
-                records.append(&mut conda_records);
-
-                // Iterate over all packages to find recursive dependencies.
-                for record in records.iter() {
-                    for dependency in &record.package_record.depends {
-                        let dependency_name = PackageName::new_unchecked(
-                            dependency.split_once(' ').unwrap_or((dependency, "")).0,
-                        );
-                        if !seen.contains(&dependency_name) {
-                            pending.push_back(dependency_name.clone());
-                            seen.insert(dependency_name);
-                        }
-                    }
-                }
-
-                result[i].append(&mut records);
-            }
-        }
-
-        Ok(result)
     }
 
     /// Returns the subdirectory from which this repodata was loaded
@@ -194,15 +95,15 @@ impl SparseRepoData {
 
 /// A serde compatible struct that only sparsely parses a repodata.json file.
 #[derive(Deserialize)]
-struct LazyRepoData<'i> {
+struct LazyMap<'i> {
     /// The channel information contained in the repodata.json file
     info: Option<ChannelInfo>,
 
-    /// The tar.bz2 packages contained in the repodata.json file
+    /// The tar.bz2 packages contained in the data file
     #[serde(borrow, deserialize_with = "deserialize_filename_and_raw_record")]
     packages: Vec<(PackageFilename<'i>, &'i RawValue)>,
 
-    /// The conda packages contained in the repodata.json file (under a different key for
+    /// The conda packages contained in the data file (under a different key for
     /// backwards compatibility with previous conda versions)
     #[serde(
         borrow,
@@ -211,80 +112,6 @@ struct LazyRepoData<'i> {
         rename = "packages.conda"
     )]
     conda_packages: Vec<(PackageFilename<'i>, &'i RawValue)>,
-}
-
-/// Parse the records for the specified package from the raw index
-fn parse_records<'i>(
-    package_name: &PackageName,
-    packages: &[(PackageFilename<'i>, &'i RawValue)],
-    base_url: Option<&str>,
-    channel: &Channel,
-    subdir: &str,
-    patch_function: Option<fn(&mut PackageRecord)>,
-) -> io::Result<Vec<RepoDataRecord>> {
-    let channel_name = channel.canonical_name();
-
-    let package_indices =
-        packages.equal_range_by(|(package, _)| package.package.cmp(package_name.as_normalized()));
-    let mut result = Vec::with_capacity(package_indices.len());
-    for (key, raw_json) in &packages[package_indices] {
-        let mut package_record: PackageRecord = serde_json::from_str(raw_json.get())?;
-        // Overwrite subdir if its empty
-        if package_record.subdir.is_empty() {
-            package_record.subdir = subdir.to_owned();
-        }
-        result.push(RepoDataRecord {
-            url: compute_package_url(
-                &channel
-                    .base_url
-                    .join(&format!("{}/", &package_record.subdir))
-                    .expect("failed determine repo_base_url"),
-                base_url,
-                key.filename,
-            ),
-            channel: channel_name.clone(),
-            package_record,
-            file_name: key.filename.to_owned(),
-        });
-    }
-
-    // Apply the patch function if one was specified
-    if let Some(patch_fn) = patch_function {
-        for record in &mut result {
-            patch_fn(&mut record.package_record);
-        }
-    }
-
-    Ok(result)
-}
-
-/// A helper function that immediately loads the records for the given packages (and their dependencies).
-/// Records for the specified packages are loaded from the repodata files.
-/// The `patch_record_fn` is applied to each record after it has been parsed and can mutate the record after
-/// it has been loaded.
-pub async fn load_repo_data_recursively(
-    repo_data_paths: impl IntoIterator<Item = (Channel, impl Into<String>, impl AsRef<Path>)>,
-    package_names: impl IntoIterator<Item = PackageName>,
-    patch_function: Option<fn(&mut PackageRecord)>,
-) -> Result<Vec<Vec<RepoDataRecord>>, io::Error> {
-    // Open the different files and memory map them to get access to their bytes. Do this in parallel.
-    let lazy_repo_data = stream::iter(repo_data_paths)
-        .map(|(channel, subdir, path)| {
-            let path = path.as_ref().to_path_buf();
-            let subdir = subdir.into();
-            tokio::task::spawn_blocking(move || {
-                SparseRepoData::new(channel, subdir, path, patch_function)
-            })
-            .unwrap_or_else(|r| match r.try_into_panic() {
-                Ok(panic) => std::panic::resume_unwind(panic),
-                Err(err) => Err(io::Error::new(io::ErrorKind::Other, err.to_string())),
-            })
-        })
-        .buffered(50)
-        .try_collect::<Vec<_>>()
-        .await?;
-
-    SparseRepoData::load_records_recursive(&lazy_repo_data, package_names, patch_function)
 }
 
 fn deserialize_filename_and_raw_record<'d, D: Deserializer<'d>>(
@@ -386,7 +213,7 @@ impl<'de> TryFrom<&'de str> for PackageFilename<'de> {
 
 #[cfg(test)]
 mod test {
-    use super::{load_repo_data_recursively, PackageFilename};
+    use super::{repodata::load_repo_data_recursively, PackageFilename};
     use rattler_conda_types::{Channel, ChannelConfig, PackageName, RepoData, RepoDataRecord};
     use rstest::rstest;
     use std::path::{Path, PathBuf};
